+++ types/node/assert.d.ts
@@ -1,7 +1,7 @@
 /**
  * The `assert` module provides a set of assertion functions for verifying
  * invariants.
- * @see [source](https://github.com/nodejs/node/blob/v16.5.0/lib/assert.js)
+ * @see [source](https://github.com/nodejs/node/blob/v16.6.0/lib/assert.js)
  */
 declare module 'assert' {
     /**
