+++ types/node/async_hooks.d.ts
@@ -6,7 +6,7 @@
  * import async_hooks from 'async_hooks';
  * ```
  * @experimental
- * @see [source](https://github.com/nodejs/node/blob/v16.5.0/lib/async_hooks.js)
+ * @see [source](https://github.com/nodejs/node/blob/v16.6.0/lib/async_hooks.js)
  */
 declare module 'async_hooks' {
     /**
