+++ types/node/base.d.ts
@@ -7,7 +7,7 @@
 //          - ~/ts3.7/index.d.ts  - Definitions specific to TypeScript 3.7 with assert pulled in
 
 // Reference required types from the default lib:
-/// <reference lib="es2018" />
+/// <reference lib="es2020" />
 /// <reference lib="esnext.asynciterable" />
 /// <reference lib="esnext.intl" />
 /// <reference lib="esnext.bigint" />
