+++ types/node/dns.d.ts
@@ -42,7 +42,7 @@
  * ```
  *
  * See the `Implementation considerations section` for more information.
- * @see [source](https://github.com/nodejs/node/blob/v16.5.0/lib/dns.js)
+ * @see [source](https://github.com/nodejs/node/blob/v16.6.0/lib/dns.js)
  */
 declare module 'dns' {
     import * as dnsPromises from 'node:dns/promises';
