+++ types/node/fs.d.ts
@@ -16,7 +16,7 @@
  *
  * All file system operations have synchronous, callback, and promise-based
  * forms, and are accessible using both CommonJS syntax and ES6 Modules (ESM).
- * @see [source](https://github.com/nodejs/node/blob/v16.5.0/lib/fs.js)
+ * @see [source](https://github.com/nodejs/node/blob/v16.6.0/lib/fs.js)
  */
 declare module 'fs' {
     import * as stream from 'node:stream';
