+++ types/node/perf_hooks.d.ts
@@ -26,7 +26,7 @@
  *   performance.measure('A to B', 'A', 'B');
  * });
  * ```
- * @see [source](https://github.com/nodejs/node/blob/v16.5.0/lib/perf_hooks.js)
+ * @see [source](https://github.com/nodejs/node/blob/v16.6.0/lib/perf_hooks.js)
  */
 declare module 'perf_hooks' {
     import { AsyncResource } from 'node:async_hooks';
