+++ types/node/scripts/generate-docs/html-doc-processing.ts
@@ -101,7 +101,7 @@ export function fixupLocalLinks(text: string, moduleName: string): string {
 export function fixupHtmlDocs(input: string, context: DocContext): string {
     // make sure code highlighting works
 
-    // This is not fast but it will do, also he is definitetely coming for us now.
+    // This is not fast but it will do, also he is definitely coming for us now.
     const hasMultiExample = /<pre><code class="language-mjs">.*?<\/code><\/pre>\n+<pre><code class="language-cjs">.*?<\/code><\/pre>/igms;
     if (hasMultiExample) {
         input = input.replace(/<pre><code class="language-cjs">.*?<\/code><\/pre>/igms, '');
