+++ types/node/ts3.6/base.d.ts
@@ -7,7 +7,7 @@
 //          - ~/ts3.6/index.d.ts  - Definitions specific to TypeScript 3.6 and earlier with assert pulled in
 
 // Reference required types from the default lib:
-/// <reference lib="es2018" />
+/// <reference lib="es2020" />
 /// <reference lib="esnext.asynciterable" />
 /// <reference lib="esnext.intl" />
 /// <reference lib="esnext.bigint" />
