+++ types/node/ts3.6/node-tests.ts
@@ -14,6 +14,7 @@ import '../test/globals';
 import '../test/http';
 import '../test/http2';
 import '../test/https';
+import '../test/inspector';
 import '../test/module';
 import '../test/net';
 import '../test/os';
@@ -25,6 +26,7 @@ import '../test/repl';
 import '../test/stream';
 import '../test/timers';
 import '../test/timers_promises';
+import '../test/trace_events';
 import '../test/string_decoder';
 import '../test/tls';
 import '../test/tty';
@@ -35,9 +37,6 @@ import '../test/vm';
 import '../test/wasi';
 import '../test/worker_threads';
 import '../test/zlib';
-
-import * as inspector from 'node:inspector';
-import * as trace_events from 'node:trace_events';
 
 /////////////////////////////////////////////////////
 /// WASI tests : https://nodejs.org/api/wasi.html ///
@@ -104,73 +103,14 @@ const importObject = { wasi_unstable: wasi.wasiImport };
  *                                                                           *
  *****************************************************************************/
 
-///////////////////////////////////////////////////////////
-/// Inspector Tests                                     ///
-///////////////////////////////////////////////////////////
-
-{
-    {
-        const b: inspector.Console.ConsoleMessage = {source: 'test', text: 'test', level: 'error' };
-        inspector.open();
-        inspector.open(0);
-        inspector.open(0, 'localhost');
-        inspector.open(0, 'localhost', true);
-        inspector.close();
-        const inspectorUrl: string | undefined = inspector.url();
-
-        const session = new inspector.Session();
-        session.connect();
-        session.disconnect();
-
-        // Unknown post method
-        session.post('A.b', { key: 'value' }, (err, params) => {});
-        // TODO: parameters are implicitly 'any' and need type annotation
-        session.post('A.b', (err: Error | null, params?: {}) => {});
-        session.post('A.b');
-        // Known post method
-        const parameter: inspector.Runtime.EvaluateParameterType = { expression: '2 + 2' };
-        session.post('Runtime.evaluate', parameter,
-            (err: Error | null, params: inspector.Runtime.EvaluateReturnType) => {});
-        session.post('Runtime.evaluate', (err: Error, params: inspector.Runtime.EvaluateReturnType) => {
-            const exceptionDetails: inspector.Runtime.ExceptionDetails = params.exceptionDetails!;
-            const resultClassName: string = params.result.className!;
-        });
-        session.post('Runtime.evaluate');
-
-        // General event
-        session.on('inspectorNotification', message => {
-            message; // $ExpectType InspectorNotification<{}>
-        });
-        // Known events
-        session.on('Debugger.paused', (message: inspector.InspectorNotification<inspector.Debugger.PausedEventDataType>) => {
-            const method: string = message.method;
-            const pauseReason: string = message.params.reason;
-        });
-        session.on('Debugger.resumed', () => {});
-    }
-}
-
-///////////////////////////////////////////////////////////
-/// Trace Events Tests                                  ///
-///////////////////////////////////////////////////////////
-
-{
-    const enabledCategories: string | undefined = trace_events.getEnabledCategories();
-    const tracing: trace_events.Tracing = trace_events.createTracing({ categories: ['node', 'v8'] });
-    const categories: string = tracing.categories;
-    const enabled: boolean = tracing.enabled;
-    tracing.enable();
-    tracing.disable();
-}
-
 ////////////////////////////////////////////////////
-/// Node.js ESNEXT Support
+/// ArrayLike.at() support
 ////////////////////////////////////////////////////
 
 {
-    const s = 'foo';
-    const s1: string = s.trimLeft();
-    const s2: string = s.trimRight();
-    const s3: string = s.trimStart();
-    const s4: string = s.trimEnd();
+    const typedArr: NodeJS.TypedArray | number[] = new Uint8Array();
+    typedArr.at(1); // $ExpectType number | undefined
+    'asd'.at(1); // $ExpectType string | undefined
+    new BigInt64Array(123).at(1); // $ExpectType bigint | undefined
+    new BigUint64Array(123).at(1); // $ExpectType bigint | undefined
 }
