+++ types/node/tsconfig.json
@@ -7,7 +7,7 @@
         "module": "commonjs",
         "target": "esnext",
         "lib": [
-            "es2018"
+            "dom"
         ],
         "noImplicitAny": true,
         "noImplicitThis": true,
