+++ types/node/v8.d.ts
@@ -4,7 +4,7 @@
  * ```js
  * const v8 = require('v8');
  * ```
- * @see [source](https://github.com/nodejs/node/blob/v16.5.0/lib/v8.js)
+ * @see [source](https://github.com/nodejs/node/blob/v16.6.0/lib/v8.js)
  */
 declare module 'v8' {
     import { Readable } from 'node:stream';
