+++ types/node/wasi.d.ts
@@ -63,7 +63,7 @@
  * The `--experimental-wasi-unstable-preview1` CLI argument is needed for this
  * example to run.
  * @experimental
- * @see [source](https://github.com/nodejs/node/blob/v16.5.0/lib/wasi.js)
+ * @see [source](https://github.com/nodejs/node/blob/v16.6.0/lib/wasi.js)
  */
 declare module 'wasi' {
     interface WASIOptions {
